extract shared CSV test fixtures into one helper module
